dqm/requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
apache-airflow==2.9.1
pyyaml==6.0.1
python-dotenv==1.0.1
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.21.0
jupyter==1.0.0
openpyxl==3.1.2
pytest==8.2.0
